wdio4: reorganize and clean up tests
