+++ packages/eyes-webdriverio-4/package.json
@@ -9,8 +9,8 @@
     "uc": "bongo --update-changelog",
     "send-release-notification": "bongo --send-release-notification",
     "lint": "eslint '**/*.js'",
-    "test": "yarn test:unit && yarn test:e2e",
-    "test:e2e": "mocha --no-timeouts \"test/+(TestClassicApi*.js|TestFluentApi*.js|TestVisualGrid*.js)\"",
+    "test": "yarn test:unit && yarn test:it && yarn test:e2e",
+    "test:e2e": "mocha --no-timeouts \"test/e2e/*.js\"",
     "test:it": "mocha --no-timeouts \"test/it/*.js\"",
     "test:unit": "mocha \"test/unit/**/*.js\"",
     "test:coverage": "coverage-tests run --path test/coverage/index.js --filterIndexes 0,1,6,7,8,9,10,11,16,17,18,27,30,31,32,33,36,39,42,60,61,78,79,80,109,114,129,130,131,137",
