+++ packages/eyes-webdriverio-4/test/e2e/TestClassicApi_Chrome_ForceFullPageScreenShot.js
