+++ packages/eyes-webdriverio-4/test/e2e/TestNativeApp.js
