+++ packages/eyes-webdriverio-4/test/e2e/TestNativeApp_Android.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const {ConsoleLogHandler} = require('@applitools/eyes-sdk-core')
-const {Eyes} = require('../index')
+const {Eyes} = require('../../index')
 const {TestNativeApp} = require('./TestNativeApp')
 const webdriverio = require('webdriverio')
 
@@ -10,7 +10,7 @@ const appName = 'NativeApp'
 let eyes = new Eyes()
 let browser
 
-describe(appName, function() {
+describe.skip(appName, function() {
   before(function() {
     eyes.setApiKey(process.env.APPLITOOLS_API_KEY)
     eyes.setLogHandler(new ConsoleLogHandler(true))
